# rpc: faster JSON generation