--- a/src/univalue/lib/univalue.cpp
+++ b/src/univalue/lib/univalue.cpp
@@ -56,20 +56,12 @@ bool UniValue::setNumStr(const std::string& val_)
 
 bool UniValue::setInt(uint64_t val_)
 {
-    std::ostringstream oss;
-
-    oss << val_;
-
-    return setNumStr(oss.str());
+    return setNumStr(std::to_string(val_));
 }
 
 bool UniValue::setInt(int64_t val_)
 {
-    std::ostringstream oss;
-
-    oss << val_;
-
-    return setNumStr(oss.str());
+    return setNumStr(std::to_string(val_));
 }
 
 bool UniValue::setFloat(double val_)
@@ -114,6 +106,16 @@ bool UniValue::push_back(const UniValue& val_)
     return true;
 }
 
+bool UniValue::push_back(UniValue&& val_)
+{
+    if (typ != VARR)
+        return false;
+
+    values.push_back(std::move(val_));
+    return true;
+}
+
+
 bool UniValue::push_backV(const std::vector<UniValue>& vec)
 {
     if (typ != VARR)
@@ -124,14 +126,28 @@ bool UniValue::push_backV(const std::vector<UniValue>& vec)
     return true;
 }
 
+void UniValue::__pushKV(const std::string& key, UniValue&& val_)
+{
+    keys.push_back(key);
+    values.push_back(std::move(val_));
+}
+void UniValue::__pushKV(std::string&& key, UniValue&& val_)
+{
+    keys.push_back(std::move(key));
+    values.push_back(std::move(val_));
+}
 void UniValue::__pushKV(const std::string& key, const UniValue& val_)
 {
     keys.push_back(key);
     values.push_back(val_);
 }
-
-bool UniValue::pushKV(const std::string& key, const UniValue& val_)
+void UniValue::__pushKV(std::string&& key, const UniValue& val_)
 {
+    keys.push_back(std::move(key));
+    values.push_back(val_);
+}
+
+bool UniValue::pushKV(const std::string& key, const UniValue& val_) {
     if (typ != VOBJ)
         return false;
 
@@ -142,7 +158,41 @@ bool UniValue::pushKV(const std::string& key, const UniValue& val_)
         __pushKV(key, val_);
     return true;
 }
+bool UniValue::pushKV(std::string&& key, const UniValue& val_) {
+    if (typ != VOBJ)
+        return false;
 
+    size_t idx;
+    if (findKey(key, idx))
+        values[idx] = val_;
+    else
+        __pushKV(std::move(key), val_);
+    return true;
+}
+bool UniValue::pushKV(const std::string& key, UniValue&& val_)
+{
+    if (typ != VOBJ)
+        return false;
+
+    size_t idx;
+    if (findKey(key, idx))
+        values[idx] = std::move(val_);
+    else
+        __pushKV(key, std::move(val_));
+    return true;
+}
+bool UniValue::pushKV(std::string&& key, UniValue&& val_)
+{
+    if (typ != VOBJ)
+        return false;
+
+    size_t idx;
+    if (findKey(key, idx))
+        values[idx] = std::move(val_);
+    else
+        __pushKV(std::move(key), std::move(val_));
+    return true;
+}
 bool UniValue::pushKVs(const UniValue& obj)
 {
     if (typ != VOBJ || obj.typ != VOBJ)
@@ -150,6 +200,17 @@ bool UniValue::pushKVs(const UniValue& obj)
 
     for (size_t i = 0; i < obj.keys.size(); i++)
         __pushKV(obj.keys[i], obj.values.at(i));
+
+    return true;
+}
+
+bool UniValue::pushKVs(UniValue&& obj)
+{
+    if (typ != VOBJ || obj.typ != VOBJ)
+        return false;
+
+    for (size_t i = 0; i < obj.keys.size(); i++)
+        __pushKV(std::move(obj.keys[i]), std::move(obj.values.at(i)));
 
     return true;
 }