--- a/src/univalue/lib/univalue_write.cpp
+++ b/src/univalue/lib/univalue_write.cpp
@@ -8,11 +8,8 @@
 #include "univalue.h"
 #include "univalue_escapes.h"
 
-static std::string json_escape(const std::string& inS)
+static void json_escape(const std::string& inS, std::string& outS)
 {
-    std::string outS;
-    outS.reserve(inS.size() * 2);
-
     for (unsigned int i = 0; i < inS.size(); i++) {
         unsigned char ch = inS[i];
         const char *escStr = escapes[ch];
@@ -22,16 +19,10 @@ static std::string json_escape(const std::string& inS)
         else
             outS += ch;
     }
-
-    return outS;
 }
 
-std::string UniValue::write(unsigned int prettyIndent,
-                            unsigned int indentLevel) const
+void UniValue::append(unsigned int prettyIndent, unsigned int indentLevel, std::string& s) const
 {
-    std::string s;
-    s.reserve(1024);
-
     unsigned int modIndent = indentLevel;
     if (modIndent == 0)
         modIndent = 1;
@@ -47,7 +38,9 @@ std::string UniValue::write(unsigned int prettyIndent,
         writeArray(prettyIndent, modIndent, s);
         break;
     case VSTR:
-        s += "\"" + json_escape(val) + "\"";
+        s += "\"";
+        json_escape(val, s);
+        s += "\"";
         break;
     case VNUM:
         s += val;
@@ -56,9 +49,16 @@ std::string UniValue::write(unsigned int prettyIndent,
         s += (val == "1" ? "true" : "false");
         break;
     }
+}
 
+std::string UniValue::write(unsigned int prettyIndent,
+                            unsigned int indentLevel) const
+{
+    std::string s;
+    append(prettyIndent, indentLevel, s);
     return s;
 }
+
 
 static void indentStr(unsigned int prettyIndent, unsigned int indentLevel, std::string& s)
 {
@@ -74,7 +74,7 @@ void UniValue::writeArray(unsigned int prettyIndent, unsigned int indentLevel, s
     for (unsigned int i = 0; i < values.size(); i++) {
         if (prettyIndent)
             indentStr(prettyIndent, indentLevel, s);
-        s += values[i].write(prettyIndent, indentLevel + 1);
+        values[i].append(prettyIndent, indentLevel + 1, s);
         if (i != (values.size() - 1)) {
             s += ",";
         }
@@ -96,10 +96,12 @@ void UniValue::writeObject(unsigned int prettyIndent, unsigned int indentLevel, 
     for (unsigned int i = 0; i < keys.size(); i++) {
         if (prettyIndent)
             indentStr(prettyIndent, indentLevel, s);
-        s += "\"" + json_escape(keys[i]) + "\":";
+        s += "\"";
+        json_escape(keys[i], s);
+        s += "\":";
         if (prettyIndent)
             s += " ";
-        s += values.at(i).write(prettyIndent, indentLevel + 1);
+        values.at(i).append(prettyIndent, indentLevel + 1, s);
         if (i != (values.size() - 1))
             s += ",";
         if (prettyIndent)