… src/bench/rpc_blockchain.cpp
… src/util/strencodings.cpp