@laanwj 2021-01-24
Nice improvement.

> Where should the univalue changes go to, into https://github.com/bitcoin-core/univalue or into https://github.com/jgarzik/univalue?

I think it's preferable if it can be merged "as upstream as possible" so that we don't have to deviate from upstream unnecessarily. E.g. in jgarzik's if possible. But if not, our repository.