@laanwj 2021-01-25
Thanks for filing it upstream. I hope it gets attention from the maintainer soon. If not, I'm happy to review it in our own fork.