@martinus 2021-01-25
I'll wait a while for jgarzik to reply, but he might not like it because this will increase dependency from c++98 to c++11